k is now 16 bits

+++ Scripts/AtA/multiway_merge/mm_utility.h
@@ -1,6 +1,6 @@
 #ifndef MULTIWAY_UTILITY_H_INCLUDED 
 #define MULTIWAY_UTILITY_H_INCLUDED 
 
-void generate_k_sorted_sublists( uint64_t *base_ptr, uint64_t total_elements, unsigned int seed, uint8_t k );
+void generate_k_sorted_sublists( uint64_t *base_ptr, uint64_t total_elements, unsigned int seed, uint16_t k );
 
 #endif // MULTIWAY_UTILITY_H_INCLUDED
